Prep A2b: move service/deploy-tool setup from builder pack → prep (settings) step
Consolidates where a user sets up services + connects deploy tools: it now lives
on the prep (settings) step next to repo connect + notifications, not on the
builder-pack page. The builder pack just reads the collected values at export.

- `components/ServiceMcpSetup.tsx` (new): self-contained — holds the services
  state, seeds via the shared `seedServiceSetup` (store-or-detect), persists every
  change to the browser-only store, and renders the service-key panel + the
  agent-aware MCP panel (with its own agent picker, since settings has no target).
- settings page: renders `<ServiceMcpSetup>` below repo connect.
- export page: dropped the two panels + their state/handlers; reads the collected
  services from the store at export time; leaves a one-line pointer
  ("서비스·배포 도구 설정은 준비 단계에서 →") so nobody who remembers the old
  location is stranded.
- `seedServiceSetup(projectId, spec)` in service-values-store: the single shared
  seed, with the two move-risk safety checks locked by tests:
    (1) values entered in prep survive to export (stored wins),
    (2) spec detection (email→Resend, …) still reaches the panel on its new screen.
- i18n `exportPage.prepMoved` (EN + KO), parity preserved.

Dashboard 498/498, typecheck + build + lint clean. Base = main.

Next: #5 (per-screen primary CTA hierarchy — highest P1).
