+++ apps/dashboard/src/lib/service-values-store.d.mts
@@ -1,3 +1,10 @@
 export declare function saveServiceValues(projectId: string, services: unknown[]): void;
 export declare function loadServiceValues(projectId: string): unknown[] | null;
 export declare function clearServiceValues(projectId: string): void;
+export declare function seedServiceSetup(
+  projectId: string,
+  spec:
+    | { oneLine?: string; problem?: string; included?: string[]; userFlow?: string[]; productName?: string }
+    | null
+    | undefined,
+): unknown[];
